frontend/.archive/requirements.txt
# FastHTML and core web framework
python-fasthtml>=0.6.0
starlette>=0.27.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0

# Database and ORM
sqlalchemy>=2.0.0
alembic>=1.12.0
aiosqlite>=0.19.0  # For async SQLite support
greenlet>=3.0.0    # Required for SQLAlchemy async

# Configuration and environment
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Authentication and security
python-multipart>=0.0.6
cryptography>=41.0.0
passlib[bcrypt]>=1.7.4

# AWS SDK integration (for Nova models)
boto3>=1.34.0
botocore>=1.34.0

# Data processing and validation
pandas>=2.0.0
numpy>=1.24.0
jsonschema>=4.19.0

# Real-time features
websockets>=11.0.0
sse-starlette>=1.6.0

# Async file operations
aiofiles>=23.0.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# === ADVANCED FEATURES DEPENDENCIES ===

# Dataset Management: Upload and manage CSV/JSON datasets with automatic processing
openpyxl>=3.1.0          # Excel file support
xlrd>=2.0.0              # Legacy Excel support
chardet>=5.2.0           # Character encoding detection
python-magic>=0.4.27     # File type detection
validators>=0.22.0       # Data validation utilities

# Prompt Engineering: Create and edit prompts with Jinja2 templating and variable detection
jinja2-time>=0.2.0       # Time extensions for Jinja2
markupsafe>=2.1.0        # Safe string handling
regex>=2023.10.0         # Advanced regex for variable detection

# Custom Metrics: Define domain-specific evaluation metrics with Python code
scipy>=1.11.0            # Scientific computing
scikit-learn>=1.3.0      # Machine learning metrics
nltk>=3.8.0              # Natural language processing
rouge-score>=0.1.2       # ROUGE metrics for text evaluation
bert-score>=0.3.13       # BERT-based semantic similarity
sentence-transformers>=2.2.0  # Sentence embeddings

# Optimization Workflows: Run automated prompt optimization with multiple algorithms
optuna>=3.4.0            # Hyperparameter optimization
hyperopt>=0.2.7          # Bayesian optimization
deap>=1.4.0              # Evolutionary algorithms
joblib>=1.3.0            # Parallel processing
tqdm>=4.66.0             # Progress bars

# AI Rubric Generation: Generate evaluation rubrics from datasets using AI
openai>=1.3.0            # OpenAI API (for comparison/fallback)
anthropic>=0.7.0         # Anthropic API (for comparison/fallback)
tiktoken>=0.5.0          # Token counting utilities

# Human Annotation: Quality assurance through human annotation workflows
redis>=5.0.0             # Session storage and task queuing
celery>=5.3.0            # Background task processing
flower>=2.0.0            # Celery monitoring
kombu>=5.3.0             # Message broker abstraction

# Real-time Progress: Live updates during optimization with WebSocket integration
python-socketio>=5.9.0   # Python Socket.IO client/server (correct package name)
eventlet>=0.33.0         # Async networking library

# Results Analysis: Comprehensive visualization and comparison of optimization results
plotly>=5.17.0           # Interactive visualizations
matplotlib>=3.8.0        # Static plotting
seaborn>=0.13.0          # Statistical visualizations
bokeh>=3.3.0             # Interactive web plots
dash>=2.14.0             # Web-based dashboards
kaleido>=0.2.1           # Static image export for Plotly

# Additional data processing and analysis
statsmodels>=0.14.0      # Statistical modeling
pyarrow>=14.0.0          # Columnar data processing
polars>=0.19.0           # Fast DataFrame library
duckdb>=0.9.0            # In-process SQL OLAP database

# File format support
python-docx>=1.1.0       # Word document processing
pypdf>=3.17.0            # PDF processing
markdown>=3.5.0          # Markdown processing
pyyaml>=6.0.0            # YAML processing
toml>=0.10.0             # TOML processing

# Monitoring and observability
prometheus-client>=0.17.0  # Metrics collection
structlog>=23.2.0        # Structured logging
sentry-sdk>=1.38.0       # Error tracking

# Performance and caching
diskcache>=5.6.0         # Disk-based caching
memory-profiler>=0.61.0  # Memory usage profiling
py-spy>=0.3.14           # Python profiler

# Security and validation
bleach>=6.1.0            # HTML sanitization
python-jose>=3.3.0       # JWT handling
argon2-cffi>=23.1.0      # Password hashing

# Development and debugging
ipython>=8.17.0          # Enhanced Python shell
jupyter>=1.0.0           # Jupyter notebook support
rich>=13.7.0             # Rich text and beautiful formatting
typer>=0.9.0             # CLI framework for admin tools

# Optional ML/AI libraries (commented out by default - uncomment as needed)
# torch>=2.1.0           # PyTorch for deep learning
# transformers>=4.35.0   # Hugging Face transformers
# datasets>=2.14.0       # Hugging Face datasets
# accelerate>=0.24.0     # Hugging Face acceleration
# langchain>=0.0.340     # LangChain framework
# chromadb>=0.4.15       # Vector database
